feat(license): Add GNU Affero General Public License
Added the GNU Affero General Public License to the project, providing users with the freedom to redistribute and modify the software under specified terms. This license ensures cooperation with the community in the case of network server software, guaranteeing users' freedom to share and change all versions of the program.

This commit adds the full text of the GNU Affero General Public License to the project, ensuring that users are aware of their rights and responsibilities when using the software.