chore(homepage info): update RAGbot and Image Upscaler descriptions
Updated the description of RAGbot and Image Upscaler in the Home.py file to reflect the current functionalities and limitations. RAGbot is inspired by AlphaCodium's paper and Image Upscaler is temporarily using Swin2SR due to resource limitations. Also added a brief explanation of the processes involved in RAGbot's functionality.